fix(project): separate project form
