+++ web/src/components/project/component/List.js
@@ -72,6 +72,7 @@ class List extends React.Component {
         title: intl.get('action'),
         dataIndex: 'metadata.name',
         key: 'action',
+        align: 'right',
         render: value => (
           <EllipsisMenu
             menuFunc={() => {
