+++ web/src/components/project/component/addProject/index.js
@@ -1,2 +1,2 @@
-import Form from './Form';
+import Form from './FormWrap';
 export default Form;
