+++ web/src/less/inc.var.less
@@ -82,22 +82,3 @@ code {
     }
   }
 }
-
-.cyclone-layout-header {
-  z-index: 200;
-  width: 100%;
-  position: fixed;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 0 20px;
-}
-
-.cyclone-logo {
-  width: 100px;
-  height: 30px;
-  line-height: 30px;
-  color: #fff;
-  font-size: 28px;
-  text-align: left;
-}
